(MERGE ERROR) Fix unintended dynamic grid resize side effects.
- Do not attempt to resize the grid unless the dynamic grid resize has
  been explicitly triggered.
- Workspace Ids are 1 indexed not 0, so on resize, start at 1.
- After resizing the grid, persist the changes to the DB. The initial
  implementation only did the resize in memory and did not persist this,
  causing side effects later.

Bugs fixed:

1. Resize the dynamic grid from large to small. Add a new icon by
installing a new app or adding a shortcut from file manager. The new
icon will overlap one on one of the new pages.

2. Resize the grid from large to small in a previous version of
trebuchet (before the grid resize feature was added). Icons on the outer
edges of the grid will be lost but persist in the DB. Upgrade to a new
version of Trebuchet. The icons will return to the workspace.

Change-Id: I8b0defafb3299d1f3d534526a75f9253495d739d